modularize app nav component - both the form and network editor use the same component for the 'more' button
