+++ src/styles/form-editor/form-editor.css
@@ -8,10 +8,11 @@
   width: 100%;
   justify-content: space-between;
   align-items: center;
+  font-size: 1.25em;
+
 }
 
 .form-editor-title {
-  font-size: 2em;
   margin-bottom: 1em;
   overflow: hidden;
 }
@@ -23,7 +24,7 @@
 }
 
 .form-home-button {
-  font-size: 2em;
+  font-size: 1.33em;
   padding-left: 0;
   margin: 0;
   border: 0;
